requirements.txt
PyGithub
openai
langchain
python-dotenv